# fix CSC, diagonal multiplication