--- a/test/linalg.jl
+++ b/test/linalg.jl
@@ -128,3 +128,9 @@ end
     sp = sprand(512, 512, 0.05)
     @test nnz(pm * sp - SparseMatrixCSC(pm) * sp) == 0
 end
+
+@testset "regression test" begin
+    pop = sparse([2, 3], [4, 5], [1, √2], 5, 5)
+    zop = Diagonal([0, 1/2, 1, -1/2, 0])
+    @test (pop * zop)[2,4] == -0.5
+end