## Previous
```julia
julia> using SparseArrays, LinearAlgebra
​
julia> pop = sparse([2, 3], [4, 5], [1, √2], 5, 5)
5×5 SparseMatrixCSC{Float64,Int64} with 2 stored entries:
  [2, 4]  =  1.0
  [3, 5]  =  1.41421
​
julia> zop = Diagonal([0, 1/2, 1, -1/2, 0])
5×5 Diagonal{Float64,Array{Float64,1}}:
 0.0   ⋅    ⋅     ⋅    ⋅ 
  ⋅   0.5   ⋅     ⋅    ⋅ 
  ⋅    ⋅   1.0    ⋅    ⋅ 
  ⋅    ⋅    ⋅   -0.5   ⋅ 
  ⋅    ⋅    ⋅     ⋅   0.0
​
julia> pop * zop
5×5 SparseMatrixCSC{Float64,Int64} with 2 stored entries:
  [2, 4]  =  -0.5
  [3, 5]  =  0.0
​
julia> using RydbergPulse
[ Info: Precompiling RydbergPulse [8e30c046-55fb-45d0-b9a3-7d9125e1d9de]
​
julia> pop * zop
5×5 SparseMatrixCSC{Float64,Int64} with 2 stored entries:
  [2, 4]  =  0.0
  [3, 5]  =  0.0
```

Now this is fixed.

## Benchmarks 
```julia
julia> using SparseArrays

julia> using LinearAlgebra

julia> pop2 = sprand(1000, 1000, 0.001);

julia> zop2 = Diagonal(randn(1000));

julia> using BenchmarkTools

julia> @benchmark $pop2 * $zop2
BenchmarkTools.Trial: 
  memory estimate:  24.75 KiB
  allocs estimate:  3
  --------------
  minimum time:     4.156 μs (0.00% GC)
  median time:      4.947 μs (0.00% GC)
  mean time:        5.352 μs (5.37% GC)
  maximum time:     91.504 μs (91.76% GC)
  --------------
  samples:          10000
  evals/sample:     7

julia> using LuxurySparse

julia> @benchmark $pop2 * $zop2
BenchmarkTools.Trial: 
  memory estimate:  24.75 KiB
  allocs estimate:  3
  --------------
  minimum time:     3.528 μs (0.00% GC)
  median time:      4.332 μs (0.00% GC)
  mean time:        4.654 μs (5.59% GC)
  maximum time:     131.605 μs (88.55% GC)
  --------------
  samples:          10000
  evals/sample:     7
```